requirements.txt
# MediLink Pharmacy Management System - Dependencies

# Web Framework
flask>=3.0.0
werkzeug>=3.0.0

# Database
pymysql>=1.1.0
cryptography>=41.0.0  # For secure MySQL connections

# Environment
python-dotenv>=1.0.0

# Security (optional but recommended for production)
flask-limiter>=3.5.0  # Rate limiting
flask-talisman>=1.1.0  # Security headers

# Testing (optional)
pytest>=7.4.0
pytest-flask>=1.3.0
